+++ src/layers/generic.h
@@ -68,18 +68,21 @@ class EncoderDecoderLayerBase : public LayerBase {
 protected:
   const std::string prefix_;
   const bool embeddingFix_;
-  const float dropout_;
+  const float dropoutEmbeddings_; // this drops out full embedding vectors 
   const bool inference_;
   const size_t batchIndex_;
   mutable std::vector<Ptr<IEmbeddingLayer>> embeddingLayers_; // (lazily created)
 
-  EncoderDecoderLayerBase(Ptr<ExpressionGraph> graph, Ptr<Options> options, const std::string& prefix, size_t batchIndex,
-        float dropout,
-        bool embeddingFix) :
+  EncoderDecoderLayerBase(Ptr<ExpressionGraph> graph, 
+                          Ptr<Options> options, 
+                          const std::string& prefix, 
+                          size_t batchIndex,
+                          float dropoutEmbeddings,
+                          bool embeddingFix) :
       LayerBase(graph, options),
       prefix_(options->get<std::string>("prefix", prefix)),
       embeddingFix_(embeddingFix),
-      dropout_(dropout),
+      dropoutEmbeddings_(dropoutEmbeddings),
       inference_(options->get<bool>("inference", false)),
       batchIndex_(options->get<size_t>("index", batchIndex)) {}
 
@@ -300,6 +303,8 @@ class Embedding : public LayerBase, public IEmbeddingLayer {
   Expr E_;
   Ptr<FactoredVocab> factoredVocab_;
   Expr multiRows(const Words& data, float dropProb) const;
+  bool inference_{false};
+
 public:
   Embedding(Ptr<ExpressionGraph> graph, Ptr<Options> options);
 
@@ -312,8 +317,11 @@ public:
 
 class ULREmbedding : public LayerBase, public IEmbeddingLayer {
   std::vector<Expr> ulrEmbeddings_; // @TODO: These could now better be written as 6 named class members
+  bool inference_{false};
+
 public:
-  ULREmbedding(Ptr<ExpressionGraph> graph, Ptr<Options> options) : LayerBase(graph, options) {
+  ULREmbedding(Ptr<ExpressionGraph> graph, Ptr<Options> options) 
+    : LayerBase(graph, options), inference_(opt<bool>("inference")) {
     std::string name = "url_embed"; //opt<std::string>("prefix");
     int dimKeys = opt<int>("dimTgtVoc");
     int dimQueries = opt<int>("dimSrcVoc");
@@ -399,7 +407,9 @@ public:
     qt = qt/sqrtDim;  // normalize accordin to embed size to avoid dot prodcut growing large in magnitude with larger embeds sizes
     auto z = dot(qt, keyEmbed, false, true);      // query-key similarity
     float dropProb = this->options_->get<float>("ulr-dropout", 0.0f);  // default no dropout
-    z = dropout(z, dropProb);
+    if(!inference_)
+      z = dropout(z, dropProb);
+
     float tau = this->options_->get<float>("ulr-softmax-temperature", 1.0f);  // default no temperature
     // temperature in softmax is to control randomness of predictions
     // high temperature Softmax outputs are more close to each other
@@ -411,7 +421,8 @@ public:
     auto graph = ulrEmbeddings_.front()->graph();
     auto batchMask = graph->constant({ dimWords, dimBatch, 1 },
                                      inits::fromVector(subBatch->mask()));
-    batchEmbeddings = dropout(batchEmbeddings, options_->get<float>("dropout", 0.0f), {batchEmbeddings->shape()[-3], 1, 1});
+    if(!inference_)
+      batchEmbeddings = dropout(batchEmbeddings, options_->get<float>("dropout-embeddings", 0.0f), {batchEmbeddings->shape()[-3], 1, 1});
     return std::make_tuple(batchEmbeddings, batchMask);
   }
 
@@ -440,7 +451,7 @@ Expr denseInline(Expr x, std::string prefix, std::string suffix, int outDim, con
   x = affine(x, W, b);
   if (actFn)
     x = actFn(x);
-  x = dropout(x, dropProb);
+  x = dropout(x, dropProb); // @TODO: check for infernce?
   return x;
 }